stopwatch: Expose function to store the whole sample
Expose function to store the whole sample, that would
be otherwise computed from start and stop timestamps,
called stopwatch_add_sample. This function is useful for
scenarios when the start would overlap, in other words
when there is no guarantee that the sequnce of start -> stop
would be called lineary.

Signed-off-by: Ales Musil <amusil@redhat.com>