test-util: Disable truncation check for Clang >= 18.
Clang is able to detect truncation errors in version 18
and above. Disable this check similarly as it was done
for GCC as the test is specifically testing the truncation.

 tests/test-util.c:1126:16: error: 'snprintf' will always be truncated; specified size is 4, but format string expands to at least 6 [-Werror,-Wformat-truncation]
 1126 |     ovs_assert(snprintf(s, 4, "abcde") == 5);
      |                ^
tests/test-util.c:1129:16: error: 'snprintf' will always be truncated; specified size is 5, but format string expands to at least 6 [-Werror,-Wformat-truncation]
 1129 |     ovs_assert(snprintf(s, 5, "abcde") == 5);
      |                ^

Signed-off-by: Ales Musil <amusil@redhat.com>