ofproto: Fix default pmd_id for ofproto/detrace.
The system and netdev datapath have different default pmd_id, which
resulted in empty output of ofproto/detrace with kernel datapath.
Also indicate that the UFID or cache wasn't available.

Make sure we use the correct default pmd_id when it's not specified
as an argument. At the same time move slightly adjusted test into
system tests so it is tested with both datapaths.

Fixes: 600125b2c380 ("ofproto: Add ofproto/detrace command to map UFIDs to OpenFlow.")
Signed-off-by: Ales Musil <amusil@redhat.com>
---
v5: Rebase on top of current main.
    Adjust the empty cache message.
    Adjust hte number of packets in the test.
    Some other smaller test fixes.
v4: Add back the group test.
    Use the OF update trick to force cache population.
v3: Rebase on top of latest main.
    Make the output work with TC offload.
    Add message for missing UFID or cache.
v2: Rebase on top of latest main.
    Address comments from Dumitru:
    - Fix typo in the commit title.
    - Simplify the mathod for pmd_id.
    - Use proper function to get dpif type.