ovsdb-idl: Add a way to assert IDL txn as read only.
Add a way to assert IDL txn as read only, that is useful in cases when
we want to ensure that the program doesn't write anything into the
txn. It is done via assert because this is considered a bug which
should be fixed in the IDL caller.

Signed-off-by: Ales Musil <amusil@redhat.com>
---
v4: Remove acks.
    Fix write to index row.

v3: Add Mike's ack.
    Address comments from Ilya:
    - Adjust the line wraps and indentation of tests.
    - Make sure we skip the test if NDEBUG is defined.

v2: Add Dumitru's ack.
    Adjust the ovsdb_idl_txn_assert_read_only() comment.
    Address some small nits.